requirement.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
antlr4-python3-runtime==4.9.3
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
Deprecated==1.3.1
diskcache==5.6.3
einops==0.8.1
filelock==3.19.1
fonttools==4.60.1
footsteps==0.1.9
frozenlist==1.8.0
fsspec==2025.9.0
girder-client==3.1.8
grpcio==1.76.0
humanize==4.14.0
hydra-core==1.3.2
icon_registration==1.1.6
idna==3.11
itk==6.0b1.post1
itk-core==6.0b1.post1
itk-filtering==6.0b1.post1
itk-io==6.0b1.post1
itk-numerics==6.0b1.post1
itk-registration==6.0b1.post1
itk-segmentation==6.0b1.post1
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lightning-utilities==0.15.2
lxml==6.0.2
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mdurl==0.1.2
meshio==5.3.5
monai==1.5.1
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
nibabel==5.3.2
nilearn==0.12.1
numpy==2.3.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.1
Pygments==2.19.2
pyparsing==3.2.5
pystrum==0.4
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.6
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
setuptools==70.2.0
shellingham==1.5.4
simpleitk==2.5.3
six==1.17.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
torch==2.8.0+cu126
torchaudio==2.8.0+cu126
torchdiffeq==0.2.5
torchio==0.21.0
torchmetrics==1.8.2
torchvision==0.23.0+cu126
tqdm==4.67.1
triton==3.4.0
typer==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
unigradicon==1.0.4
urllib3==2.5.0
vedo==2025.5.4
vtk==9.5.2
Werkzeug==3.1.3
wrapt==2.0.1
yarl==1.22.0
